AzureFunctions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.32.0
azure-functions==1.21.3
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==3.3.2
idna==3.10
isodate==0.7.2
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0